// src/types/product.d.ts
export interface Product {
  application: string,
  salePrice: number,
  manufacturerName: string,
  hardness: number,
  grossPrice: number,
  BUPUnitOfMeasure: string,
  articleNumber: string,
  stepQuantity: number,
  BUPValue: string,
  badges: string[],
  unitOfMeasure: string,
  color: string,
  productImage: string,
  temperatureRange: string,
  BUPConversion: string,
  minQuantity: number,
  manufacturerImage: string,
  name: string,
  material: string,
  sku: string,
  cordThickness: number,
  internalDiameter: number,
  sizeAS568: number,
  listPrice: number,
  channel: string,
  display: boolean,
  atp: {}
}